(MAINT) Improve authorization target handling (#8920)
Prior to this change, the parameter handling for the authorization action
was implemented with a minimal handler for the **target** input and the
associated workflow used an incorrect list of file globs to check for.

This change corrects the file globs for the workflow and updates the
implementation of the action to handle targets separated by commas and
newlines, including when those separators are intermingled. It also ensures
that the items in the target list passed to the action script are trimmed
of whitespace and surrounding quotes.

This enables a more maintainable workflow file for consumers of the action
by allowing them to use multiple lines to list their targets, making the
use of commas optional when specifying targets on multiple lines, and
handling indentation and quoting. This implementation still functions for
simpler single-line declarations of the input and is backwards-compatible.