(AB-1951698) Add Changed Content Version Reporting (#8931)
When managing documentation with multiple versions of the same file, it
is necessary to check Pull Requests modifying versioned content across
all of the versions that file exists for. It is not always obvious from
GitHub's view for files changed in a Pull Request whether and how a file
was modified across versions.

Prior to this change, the only way for the maintainers to inspect this
information was to carefully read every change in the Pull Request. It
required a large amount of cognitive load as a reviewer needed to make a
mental note of every change to versioned content, find the other
versions of that file (if any) the Pull Request modified, and verify
whether all needed changes were made.

This change helps to address the problem by introducing a new GitHub
Action for reporting on how a Pull Request modified versioned content.

The new `reporting/versioned-content` action inspects the files changed
in a Pull Request to determine which files are versioned content and
then writes a report by folder enumerating which files the Pull Request
changed and how those changes impact each version for the file.

The action is called in this repository by the new `versioned-content`
workflow, which runs on Pull Requests to the `main` branch which modify
versioned content in the `reference` folder.

Like the other actions, `reporting/versioned-content` can be called from
other repositories.

This change also added new utility functions to the GHA module:

- `Add-VersionedContentTable` writes a markdown table summarizing a
  changeset.
- `Get-VersionedContentChangeStatus` analyzes a Pull Request's changes
  to return information about the changes to versioned content.
- `Get-VersionedContentTableColumnWidth` returns the calculated widths
  for columns in the report's tables to ensure the pipes in the tables
  line up for reading when interactively debugging.

All three functions are exported by the module.

Finally, the action utilizes the updated `Get-VersionedContentReport`
script, which is now appropriately documented.

- Resolves AB#1951698