(GH-8952) Update VS Code docs on version selection
Prior to this change, the documentation for selecting a version of
PowerShell to use with the extension was incorrect due to a change
in the VS Code APIs that altered how languages can use the status
bar.

This change updates the language in the documentation to reflect
the current behavior/UI and explains how the old UI can be regained
by manually pinning the version to the status bar.

This change also updates the screenshots for the other articles,
which showed the old UI.