Improve consistency of documentation relevant to installation onto Red Hat Enterprise Linux. (#9126)
If the command assumes positive response during invocation of `yum`, it should when it invokes `dnf`. This is supported not least by how for recent versions of RHEL and Fedora, `yum` is a merely symbolic link to `dnf` anyway.

Additionally, I have utilized the long versions of the flags so that the documentation is easier for users ignorant of what the flags perform to comprehend what they perform. Causes installation to be less scary.