(AB-4196) Codify markdownlint (#9131)
Prior to this change, the repository had a markdownlint configuration.
This change updates that configuration, adds a minimal configuration for
markdownlint-cl2 (which the VS Code extension uses), adds a markdownlint
configuration to the about folders for reference documentation, and adds
a markdownlint ignore directive to the `includes` files.

For the root-folder configuration, this change:

- Adds line length enforcement for code blocks at 90 characters and
  turns the rule on for code blocks. Previously, it was disabled.
- Adds three new rules, all enabled. These rules are not yet picked up
  by the VS Code extension, but will be in the next release.

  - `link-fragments` requires that links to anchors are valid.
  - `reference-links-images` requires that reference-style links use a
    defined label (the `[foo]: https//...` text at the bottom of a doc).
  - `link-image-reference-definitions` requires that definedd labels for
    reference-style links are used at least once.

The root-folder `.markdownlint-cli2.yaml` file governs the behavior of
the linter overall. The settings included are minimal and do not effect
the editing experience. This file may require iteration for advanced
usage later.

In every include file, this change adds a markdownlint directive to
ignore the `first-line-h1` rule. These files are always included in the
body of another article and should not start with an H1.

In every `About` folder for reference content, this change adds a
`.markdownlint.json` file. When editing Markdown in VS Code, the editor
uses the closest configuration file to the edited document. That allows
us to override and specify alternate rules on a per-folder basis.

In this change, the new configuration files in the `About` folders:

- Extend the root configuration. Any settings not specified in these
  configurations is inherited from the configuration at the project
  root.
- Overrides the `line-length` rule, setting the maximum line length for
  code blocks to 74 characters and for other text to 79 characters. This
  is to accomodate the PowerShell Help System's line length
  requirements.

This change is related to AB#4196 and AB#4197.