(AB#5132) Add primary cSpell configuration
This change replaces the minimal cSpell configuration with a
fuller configuration in YAML for the PowerShell documentation
and adds cSpell to the list of suggested extensions.

It includes a basic dictionary for PowerShell docs, one for
PowerShell aliases, one for file extensions, one for fictional
corporations, and one for external commands.

This configuration should be copied to the other repositories
and _extended_ as necessary. Changes to these dictionaries
should be made in this repository, not the downstream ones.