(GH-9194) Document `$LastExitCode` behavior for scripts
Prior to this change, the documentation for `$LastExitCode` in
`about_Automatic_Variables` specified that the value was set only by
the last native program (for PowerShell) or last Windows-based program
(for Windows PowerShell) that was run.

As noted in #9194, this is inaccurate, as the `exit` keyword can be used
to set this value when a script is called.

This change:

- Cleans up syntax/prose for maintainability
- Clarifies that `$LastExitCode` can be populated by a PowerShell script
- Clarifies how calling a PowerShell script affects `$LastExitCode`
- Fixes #9194
- Fixes AB#5371