(MAINT) Use pull_request_target for checklist
Prior to this change, the workflow for verifying the PR checklist ran on
`pull_request`. When the Azure Boards bot edits a PR body to associate
it with an Azure DevOps item, the last `edited` event that triggers the
checklist workflow is from the bot, not a user. Because the bot has
never contributed to the project in a way GitHub understands---it only
edits comments, never commits code or files an issue---it always trips
the GitHub configuration for requiring approval for first-time
contributors for PRs from forks.

This change is an attempt to work around this limitation by moving the
checklist workflow from `pull_request` to `pull_request_target`, which
should run in the context of the upstream repository.