(MAINT) Fix hacktoberfest automation
Prior to this change, the logic for the hacktoberfest GitHub Actions was
broken, firing on every PR and issue even when they didn't match. This
was due to the outputs of the check jobs being strings, not booleans,
and the `if` statements in the dependent jobs not explicitly checking
for true/false against the string value.

This change ensures the automation runs only when expected.