(GH-9357) Clarify SessionOption for Invoke-Command (#9361)
* (GH-9357) Clarify SessionOption for Invoke-Command

Prior to this change, Example 18 for `Invoke-Command` raised an error
because the **IdleTimeout** value wasn't high enough. The root of this
problem is that the example was casting a value for milliseconds to
ticks. The **SessionOption** parameter documentation also wasn't
explicit that specifying a hashtable for the parameter is different
from calling `New-PSSessionObject`.

This change:

- Updates example 18 to use `[timespan]::FromHours()` to make the
  example functional and easier to understand
- Adds a note to the **SessionOption** parameter documentation about
  the behavior when a user specifies a hashtable
- Fixes Vale and markdownlint warnings where possible
- Standardizes example code snippets for readability
- Fixes #9357
- Fixes AB#26713

* Apply suggestions from code review

Co-authored-by: Sean Wheeler <sean.wheeler@microsoft.com>