(AB#30882) Configure Quality Contributions project
Prior to this change, the repository was configured with issue templates
and workflows to support the Hacktoberfest 2022 effort and GitHub
project.

This change builds on the infrastructure and lessons learned to create
a quality contributions project. This project is for encouraging,
supporting, and tracking community contributions to improving the
quality of PowerShell documentation across Learn.

This change:

- Includes a new page in the contributor's guide with instructions for
  participating
- Replaces the Hacktoberfest issue template with a Quality Contribution
  template and updated information
- Replaces the Hacktoberfest entry in the issue picker with a link to
  the new page in the contributor's guide
- Reworks and renames the Hacktoberfest workflows for the project
- Fixes AB#30882