(AB#4185) Explain GH labels in contributor's guide
Prior to this change, the contributor's guide included references to the
labels we use for tracking and categorizing work in GitHub but didn't
include specific information about them.

While the labels do have descriptions in GitHub, the length of the
descriptions is limited and only plain text is allowed.

This change introduces a new document for the contributor's guide that
explains all of the GitHub labels, providing additional context and
links where needed.

- Fixes AB#4185