(GH-9479) Clarify delay-binding
Prior to this change, the documentation in `about_Script_Blocks` showed
how to use delay-binding but didn't explain how they're useful and
wasn't clear that this is a usability feature for cmdlet users, not
for cmdlet authors.

This change adds some clarification and extends the example to show how
delay-binding can help users write shorter, more convenient code.

- Resolves #9479
- Fixes AB#43576