(GH-9480) Update validation attribute & examples
Prior to this change, the documentation for validation attributes in the
about topic `about_Functions_Advanced_Parameters` stated:

> When you use a type converter along with a validation attribute,
> the type converter has to be defined before the attribute

This isn't true and several of the examples don't follow this claim.
The origin of this note appears to be a workaround for using the
**ValidateLength** attribute with an integer parameter or variable.

With the attribute to the left of the type, the statement fails because
an integer doesn't have a length and the attribute can't be satisfied.
With the attribute to the right of the type, the first assignment can be
successful, but you can't assign values to the variable afterward.

This is usage is an anti-pattern that shouldn't be recommended, used as
default advice, or shown since it's not only an anti-pattern but also
a confusing and non-obvious usage.

This change:

- Replaces the direction for placing attributes to the right of types
  with direction for placing attributes to the left of types with
  supporting reasoning and examples
- Updates the examples to show coherent usages for each attribute
- Corrects the **ValidateScript** example to actually pass
- Resolves #9480
- Fixes AB#43577