(GH-9521) Add `about_PSItem` article
Prior to this change, information on the `$PSItem` variable and its
alias were only partially covered in `about_Automatic_Variables`. It was
referenced elsewhere with more examples but not in a very discoverable
way.

This change addresses that issue by adding a new `about_PSItem` article
that collates use cases into one place with examples.

- Resolves #9521
- Fixes AB#47751