(GH-9744) Fix about_Using examples (#9754)
Prior to this change, examples 2 and 3 for about_Using
were misleading. Example 2 implied that the classes from
a script module could be namespaced. Example 3 implied
that authors could call `using` to use classes from the
assembly when defining a class in PowerShell, which
actually raises a parsing exception.

This change reworks both examples to function without
errors or incorrect implications.

- Resolves #9744
- Fixes AB#60738