(AB#61939) Add devcontainer
This change adds a devcontainer definition to the repository. The
devcontainer can be used locally or as the backing definition for
a codespace and includes:

- PowerShell
- Git
- The GitHub CLI
- Vale
- All suggested VS Code extensions for the repository

This change enables us to move forward with advising contributors
to use GitHub Codespaces for their contributions to reduce the
friction for setting up their editor.

- Resolves AB#61939