Write-Host is a wrapper for Write-Information, it should be listed as a Write Cmdlet on Stream 6 (#9821)
* Update about_Redirection.md

Starting in Windows PowerShell 5.0, `Write-Host` is a wrapper for `Write-Information` This allows you to use `Write-Host` to emit output to the information stream. This enables the capture or suppression of data written using `Write-Host` while preserving backwards compatibility.

Ref: https://learn.microsoft.com/en-us/powershell/module/microsoft.powershell.utility/write-host?view=powershell-5.1#description

* Update about_Redirection.md

Starting in Windows PowerShell 5.0, `Write-Host` is a wrapper for `Write-Information` This allows you to use `Write-Host` to emit output to the information stream. This enables the capture or suppression of data written using `Write-Host` while preserving backwards compatibility.

Ref: https://learn.microsoft.com/en-us/powershell/module/microsoft.powershell.utility/write-host?view=powershell-7.2

* Update about_Redirection.md

Starting in Windows PowerShell 5.0, `Write-Host` is a wrapper for `Write-Information` This allows you to use `Write-Host` to emit output to the information stream. This enables the capture or suppression of data written using `Write-Host` while preserving backwards compatibility.

Ref: https://learn.microsoft.com/en-us/powershell/module/microsoft.powershell.utility/write-host?view=powershell-7.3#description

* Update about_Redirection.md

Starting in Windows PowerShell 5.0, `Write-Host` is a wrapper for `Write-Information` This allows you to use `Write-Host` to emit output to the information stream. This enables the capture or suppression of data written using `Write-Host` while preserving backwards compatibility.

Ref: https://learn.microsoft.com/en-us/powershell/module/microsoft.powershell.utility/write-host?view=powershell-7.4#description