Document improvements to web cmdlets for 7.4 (#9903)
* (GH-9689) Add AllowInsecureRedirect to web cmdlets

This change adds the **AllowInsecureRedirect** parameter to
the web cmdlets.

- Fixes AB#70151
- Resolves #9689

* (GH-9690) Add PreserveHttpMethodOnRedirect to web cmdlets

This change adds the **PreserveHttpMethodOnRedirect** parameter
to the web cmdlets.

- Fixes AB#70152
- Resolves #9690

* (GH-9851) Update docs for ContentType in web cmdlets

This change documents the updated behavior for preferring the value of
the **ContentType** parameter when the `Content-Type` header is defined
in the value of the **Headers** parameter.

- Fixes AB#70154
- Resolves #9851

* (GH-9863) Document PUT behavior for ContentType

This change updates the docs for 7.4 to include the changed behavior
for automatically setting the content type for PUT requests when the
content type is not specified, aligning the behavior to the existing
behavior for POST.

- Fixes AB#70155
- Resolves #9863

---------

Co-authored-by: Sean Wheeler <sean.wheeler@microsoft.com>