(MAINT) Add Vale config to repo
Prior to this change, using Vale to lint the docs
required knowledge of the tool and where to
retrieve the configuration and rules.

This change adds a new `.vale.ini` file to the
repository root and configures it to use our own
published prose style package.

It also ensures the `.vale` and `.vscode/styles`
folders are ignored by git for contributors who
use `Install-WorkspaceVale` and sync the style
packages respectively.