(GH-9948) Fix AddToHistoryHandler in Set-PSReadLineOption
This change corrects the documentation for the **AddToHistoryHandler**
parameter of the `Set-PSReadLineOption` cmdlet. Prior to this change,
the documentation stated that the valid return values were true/false.

The valid return values also include the **AddToHistoryOption** enum
and strings matching the names of those enums.

This change corrects the documentation and further explains how each
value affects the history.

- Resolves #9948
- Fixes AB#69559