(GH-9987) Clarify return behavior for collections with `-like`
Prior to this change, the documentation for the `-like` and `-notlike` operator
did not cover the (potentially surprising) behavior when called for a collection
where the members of the collection are not strings.

In this case, the operators return the matching items from the collection cast
to `[string]`, unlike the `-match` and `-notmatch` operators, which return the
input object.

This change:

- Clarifies the behavior for these operators
- Resolves #9987
- Fixes AB#83269