(GH-9997) Recommand **List<T>** over array addition
Prior to this change, the array addition section of the article
_PowerShell scripting performance considerations_ recommended that
script authors use the **ArrayList** type instead of using array
addition to add new values to an array.

The use of **ArrayList** is discouraged in the [.NET documentation][01],
recommending instead that developers use **List<T>** with a specific
type for homogenous lists or **List<Object>** for heterogenous lists.

This change:

- Recommends using **List<T>** instead of array addition or **ArrayList**
- Adds performance comparisons to clarify the impact of array addition
- Resolves #9997
- Fixes AB#84476

[01]: https://learn.microsoft.com/en-us/dotnet/api/system.collections.arraylist?view=net-8.0#remarks