(GH-10003) Clarify OutputType notes
Prior to this change, the notes section of the `about_Functions_OutputTypeAttribute`
topic referenced the `Get-Time` function from a prior example without connecting the
note to the example. The example could instead use the `Get-Date` cmdlet, which is
available and will work regardless of whether the user has run prior examples in
their terminal.

This change:

- Updates the notes section to use `Get-Date` instead of `Get-Time`.
- Cleans up linting errors and warnings.
- Ensures correct formatting for types and attributes.
- Resolves #10003
- Fixes AB#85082