(GH-10076) Remove incorrect info from `New-LocalUser`
Prior to this change, the reference documentation for the `New-LocalUser`
cmdlet incorrectly stated that it can be used to create a local account that is
connected to a Microsoft account.

However, SAM account names, and any API call for creating user accounts, is
[limited to 20 characters][01]. This problem was identified in #997 and
partially addressed by removing an incorrect example.

With the 20 character limit, a user can't specify a valid email after the
`MicrosoftAccount\` prefix - the smallest user name,
`MicrosoftAccount\a@outlook.com`, is 30 characters long.

This change removes the statement about the cmdlet being able to create
local accounts connected to a Microsoft account.

[01]: https://learn.microsoft.com/en-us/windows/win32/adschema/a-samaccountname