(GH-10459) Reorganize grouping expression docs
Prior to this change, the documentation for using grouped expressions
in _about_Operators_ included two subheadings:

1. Grouping assignment statements
1. Piping grouped expressions

However, it also defined a third use-case for grouping, which showed
how to use the grouping operator to use the output of a command in
an expression. This use case was documented before the subheadings and
after a sentence indicating there's ways to use the grouping operator
beyond overriding operator precedence.

Further, the section for grouping assignment statements was very long
compared to the other use-cases and placed in the middle, despite being
less commonly used than the others.

This change:

- Adds a subheading for the first use case to make it equivalent to the
  others in the information hierarchy.
- Moves the section for piping grouped expressions to before the section
  for grouping assignment statements.
- Resolves #10459
- Fixes AB#163845