(GH-10505) Add note on property name ordering
Prior to this change, the "Enumerating property names" section
in _Everything you wanted to know about PSCustomObject_ didn't
clarify that the results from using `Get-Member` and the
member-access operator to enumerate property names differs.

`Get-Member` returns the properties in alphabetical order while
member-access returns them in the order they were defined on the
object.

This change:

- adds a note to clarify this difference.
- Resolves #10505
- Fixes AB#169145