(AB-168933) Clarify Length and Count for PSCustomObject
Prior to this change, the documentation note for the **Length**
and **Count** properties of **PSCustomObjects** created by
casting a hashtable indicated that, for Windows PowerShell,
those properties don't exist on the object and accessing them
returns `$null`.

However, in PowerShell, those members exist and return a value
of `1`.

This change:

- Updates the note to indicate the changed behavior.
- Fixes AB#168933