(MAINT) Clarify automatic vars in classes
Prior to this change, the class documentation did not clearly explain
which automatic variables are available in class methods other than
the `$this` variable. This change adds the list of available automatic
variables and how to access them.

It also makes minor corrections to links and uses `Update-TypeData`
instead of `Add-Member` consistently. Finally, it explains examples
that were previously missing any information.