Update New-Alias.md to fix typo, improve clarity, and add new example (#11150)
* Update New-Alias.md to fix typo, improve clarity, and add new example

I found a typo in the Example 2 explanation. Also, since I was adjusting that text, I modified the text for improved clarity. Also, I added Example 3 to show how to use a custom function with an alias since the New-Alias cmdlet will not take a Value other than a cmdlet or an executable in the form of a string value.

* Editorial updates and copy changes to all versions

---------

Co-authored-by: Sean Wheeler <sean.wheeler@microsoft.com>