(AB#263698) Fix broken anchor in `Compare-Object` docs
Prior to this change, the document used a self-closing
anchor tag to enable quick-linking to the third example.

This change updates the anchor to wrap the text `Example 3`
in the heading, per guidance for the platform. This ensures
that the remainder of the page isn't a link to the anchor.

- Fixes AB#263698