(GH-11534) Separate codeblocks for installing with WinGet by ID (#11535)
Prior to this change, the article for installing PowerShell on Windows used a
single codeblock to show readers how to install PowerShell and PowerShell preview
with WinGet by ID. On the webpage, this results in a single codeblock with a copy
button. If a reader uses the copy button and pastes the copied code into their
terminal, both commands will run. If a user wanted to install only one edition,
they need to edit the copied code to remove the line they don't want.

This change:

- Separates the codeblocks for installing PowerShell with WinGet and specifying
  the ID parameter for PowerShell and PowerShell preview, so that readers may
  copy either codeblock and paste into their terminal instead of copying both.
- Fixes #11534
- Fixes AB#343802