Fix inaccuracies in `about_Language_Modes` and minor formatting changes (#11979)
* Remove invalid types from v5.1's core type list

The types removed from the core types list (permitted in
ConstrainedLanguage mode) are only applicable to PS v6+, so are removed
from the Windows PowerShell v5.1 document.

Unlike in later versions, [ordered] is not a type (accelerator) in
v5.1.

* Add [ref] special case for ConstrainedLanguage

[ref] is found in CoreTypes and is therefore permitted in
ConstrainedLanguage mode. However, casting is not permitted and results
in a statement-terminating error. Other uses (new() instantiation, in
parameter binding, etc) are permitted.

* Minor formatting changes to about_Language_Modes

* Add note on Windows-only CoreTypes

* Add note on [ordered] caveats in v5.1 CL mode

In v5.1, [ordered] is not a type, nor is OrderedDictionary included in
CoreTypes. [ordered] @{} syntax is permitted in v5.1
ConstrainedLanguage mode, but use of the object is limited.

* Remove invalid variable from v5.1's RL mode list