Improve Get-Process Ex 4 Priority grouping
The Priority table view for Diagnostics.Process groups on the
PriorityClass property. Format-Table expects input to already be
sorted, but the original example used unsorted data, rendering it
fairly unhelpful.

This adds a call to Sort-Object and sorts on PriorityClass before
calling Format-Table. This ensures all processes are correctly grouped
by the Priority view. A script block is used in lieu of the property
name directly, as the latter surfaces getter exceptions as a
non-terminating error. This will occur with, e.g., PPL processes in
Windows (even when elevated).

Reference to PS home directory .ps1xml files is also removed from the
v7.x docs. Format definitions were moved directly into source code in
v6+.

As a Format-* cmdlet is used, the term "display" is used to emphasize
output is not intended for further processing.