(GHA) Update the auth action to accept an allowlist (#12181)
Prior to this change, the verification/`authorization` GitHub Action
only supported checking the assigned permissions for a user. This
worked for normal accounts. However, the managed bot account for the
Learn platform doesn't have permissions for this repository.

This change adds a new (backwards-compatible) `authorized_accounts`
parameter to the GHA. Repository maintainers can now define an
allowlist to use for authorization in addition to the permissions
to check. If a user is explicitly in the allowlist, the action skips
checking their permissions. If a user isn't in the allowlist, they
can still pass authorization if they have matching permissions.