(GHA) Fix authorization workflow (#12252)
Prior to this change, the authorization workflow used the PowerShell
script parameter names and casing for the action parameter names.
This worked for all previous parameters because PowerShell is case
insensitive and the parameters were only one word.

However, in #12181, we added the `authorized_accounts` parameter
to the workflow. For GitHub Actions, we use `snake_case` for
parameter names, where PowerShell best practice is to use
`PascalCase`.

PowerShell's case insensitivity meant that we were able to handle
the prior parameters, which the workflow sends to the handler
scripts as the `INPUT_<ParameterName>` environment variables.

When we added `authorized_accounts`, the handler scripts were
looking for `INPUT_AuthorizedAccounts`, which doesn't exist.

This change updates the `Name` key for every defined parameter in
the `verification/authorization/v1` workflow to match the actual
workflow parameter name and casing.