Update about_CommonParameters.md: Pipeline variable example (#12392)
* Update about_CommonParameters.md: Pipeline variable example

Following a string of questions by a PowerShell learner, I realized that the `PipelineVariable` example in this document is counter-intuitive. So, I wrote a better one. I'll include the details in the PR message.

* Update about_CommonParameters.md: Pipeline variable example

Following a series of questions by a PowerShell learner, I realized that the `PipelineVariable` example in this document is counter-intuitive. So, I wrote a better one. I'll include the details in the PR message.

* Update about_CommonParameters.md: Pipeline variable example

Following a series of questions by a PowerShell learner, I realized that the `PipelineVariable` example in this document is counter-intuitive. So, I wrote a better one. I'll include the details in the PR message.

* Edits for formatting and improving prose

---------

Co-authored-by: Sean Wheeler <sean.wheeler@microsoft.com>