Add -and/-or precedence example
This adds an example involving both -and and -or to
about_Operator_Precedence. It demonstrates the equal precedence
between the two operators and calls out how this differs from other
languages such as C#.

It also notes that other contexts *within* PowerShell such as WQL and
the AD filter have their own operator precedence which might differ
(in both aforementioned cases, AND has higher precedence than OR).